chore(deps)!: upgrade Astro 5 -> 7 and Vite 6 -> 8
Two major jumps, and Astro 6 removes the legacy content collections API this
site was built on, so this is a migration rather than a version bump.

Dependencies:
- astro 5.18.0 -> 7.2.1
- vite 6.4.1 -> 8.2.1 (Astro 7 bundles Vite 8; the two move together)
- adds @astrojs/markdown-remark and zod as direct dependencies

Content collections (the actual work). Two legacy behaviours were load-bearing
for this site and both disappear with the legacy API:

- `slug` frontmatter overrode the path-derived slug, which is how posts filed
  under YYYY/MM/ get served flat at /blog/<slug>/. The content layer has no
  reserved `slug` key, so the glob loader's `generateId` rebuilds it. Without
  this every post URL would gain a /YYYY/MM/ segment.
- `layout` frontmatter was applied automatically. It no longer is, so
  blog/[slug].astro imports BlogPost.astro and wraps <Content /> itself.

Also: entries are now addressed by `id` rather than `slug`, `entry.render()`
became `render(entry)`, `src/content/config.ts` moved to `src/content.config.ts`,
and `z` now comes from zod directly since Astro 7 deprecated the `astro:content`
re-export and no longer ships it as a namespace.

Post ordering needed an explicit fix. Legacy collections returned entries in
source-path order, so a bare `.reverse()` meant "newest first" only because
filenames are date-prefixed. The content layer orders by `id` — the flat slug —
which made `.reverse()` reverse-alphabetical. Added `getPostsNewestFirst()` in
`@utils/posts` and pointed the three call sites at it.

Two Astro 7 defaults are overridden to avoid changing rendered output as a side
effect of a dependency bump: `compressHTML` stays `true` (the new 'jsx' default
strips whitespace between inline elements) and the Markdown processor stays
remark/rehype rather than the new Sätteri default, since the posts came out of
Ghost and shikiConfig feeds that pipeline.

Verified by diffing all 215 built pages against an Astro 5 build:
- identical route set, identical image set, identical RSS links and sitemap
- visible text identical on 214/215 pages
- the one text difference is the blog index, where two posts sharing 2023-04-23
  swap: sorting by pubDate puts the 4:13 PM post ahead of the 9:36 AM one, which
  filename ordering had backwards
- one heading anchor changed, #necessary-backstory -> #necessary-backstory-, from
  a slugger change in the newer markdown-remark. The heading id and its table-of-
  contents link move together, so in-page navigation is unaffected; only an
  external deep link to that anchor would drift.

`pnpm run build` emits 215 pages and `pnpm run check` is clean. Dev server,
OG image generation, and page rendering all spot-checked in a browser.
