+++ src/components/generic/RecentBlogPosts.astro
@@ -1,5 +1,5 @@
 ---
-import { getCollection } from 'astro:content';
+import { getPostsNewestFirst } from '@utils/posts';
 import BlogList from '../blog/BlogList.astro';
 import { Card } from '@eliancodes/brutal-ui';
 import { Button } from '@eliancodes/brutal-ui';
@@ -10,7 +10,7 @@ interface Props {
 
 const { count } = Astro.props;
 
-const posts = await getCollection('blog').then((posts) => posts.reverse().slice(0, count ?? 6));
+const posts = await getPostsNewestFirst().then((posts) => posts.slice(0, count ?? 6));
 ---
 
 <section class="mt-8">
