+++ src/pages/blog/index.astro
@@ -1,9 +1,9 @@
 ---
-import { getCollection } from 'astro:content';
 import Layout from '@layouts/Default.astro';
 import BlogList from '@components/blog/BlogList.astro';
+import { getPostsNewestFirst } from '@utils/posts';
 
-const posts = await getCollection('blog').then((collection) => collection.reverse());
+const posts = await getPostsNewestFirst();
 ---
 
 <Layout
