+++ src/pages/feed.xml.js
@@ -13,7 +13,7 @@ export async function GET(context) {
       title: post.data.title,
       pubDate: post.data.pubDate,
       description: post.data.description,
-      link: `/blog/${post.slug}/`,
+      link: `/blog/${post.id}/`,
     })),
     customData: '<language>en-us</language>',
     canonicalUrl: 'https://ephbaum.dev',
