+++ src/pages/v1/generate/og/[slug].png.ts
@@ -43,7 +43,7 @@ const posts = await getCollection('blog');
 
 export function getStaticPaths() {
   return posts.map((post) => ({
-    params: { slug: post.slug },
+    params: { slug: post.id },
     props: { title: post.data.title, description: post.data.description },
   }));
 }
